requirements.txt
datasets
matplotlib
numpy
pandas
Pillow
pytesseract
scikit_learn
tqdm
transformers
